backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pillow==10.0.1
pdf2docx==0.5.6
python-docx==1.0.1
docx2pdf==0.1.8
moviepy==1.0.3
pydub==0.25.1
pytesseract==0.3.10
pypdf2==3.0.1
python-magic==0.4.27
aiofiles==23.2.1
ffmpeg-python==0.2.0
zipfile36==0.1.3
rarfile==4.0
py7zr==0.20.5
opencv-python==4.8.1.78
numpy==1.26.0
markdown==3.5.1
html2text==2020.1.16
beautifulsoup4==4.12.2
pyyaml==6.0.1
xmltodict==0.13.0
dicttoxml==1.7.16
pdfkit==1.0.0
reportlab==4.0.7
cairosvg==2.7.1
